Fixed trailing slashes and relative URLs.
I learned today that adding a "/?" to the end of your routes in Sinatra
will prevent users from being 404'd when they enter something with or
without a trailing slash that doesn't have an explicit matching route.
This really should be default behavior in my opinion (or at least used
throughout the documentation), but it's not.

I also tried to deploy this application to a subdirectory using
RackBaseURI and Apache + mod_passenger, which broke everything.  It
turns out that Sinatra has a URL helper now to fix this automatically
for you ("url", aliased "to").  I've added it to every link in the
application, which should allow it to work without modification
regardless of how it's being hosted.